requirements.txt
fastapi
uvicorn[standard]
gunicorn
line-bot-sdk
requests
beautifulsoup4
pillow